full integration is fixed

+++ code/report_inserter.py
@@ -16,13 +16,15 @@ class ReportInserter:
     target_folder = "./output/relations/{}"
     iterations = 0
     parse_txt = True
+    first = False
 
     def __init__(self):
         self.db = Database()
 
-        self.db.drop_table("reports")
-        self.db.drop_table("specifies")
-        self.db.drop_table("report")
+        if self.first:
+            self.db.drop_table("reports")
+            self.db.drop_table("specifies")
+            self.db.drop_table("report")
 
         Helper.create_tables("./sql/create_report.sql", self.db)
         Helper.create_tables("./sql/create_specifies.sql", self.db)
@@ -44,13 +46,14 @@ class ReportInserter:
             logging.debug("finished parsing all csv after {}s".format(round(time.time() - self.start), 2))
 
         self.start = time.time()
-        self.db.load_infile(self.target_folder.format("report.txt"), "Report")
-        logging.debug("finished loading report after {}s".format(round(time.time() - self.start), 2))
+        if self.first:
+            self.db.load_infile(self.target_folder.format("report.txt"), "Report")
+            logging.debug("finished loading report after {}s".format(round(time.time() - self.start), 2))
 
-        self.db.load_infile(self.target_folder.format("reports.txt"), "reports")
-        logging.debug("finished loading reports csv after {}s".format(round(time.time() - self.start), 2))
+            self.db.load_infile(self.target_folder.format("reports.txt"), "reports")
+            logging.debug("finished loading reports csv after {}s".format(round(time.time() - self.start), 2))
 
-        #self.db.load_infile(self.target_folder.format("specifies.txt"), "specifies")
+            #self.db.load_infile(self.target_folder.format("specifies.txt"), "specifies")
         logging.debug("finished loading all csv after {}s".format(round(time.time() - self.start), 2))
 
     def parse_cols(self, cols):
@@ -73,10 +76,12 @@ class ReportInserter:
 
         logging.debug("Finished sorting chunk {} in {}s".format(self.iterations, round(time.time() - self.time, 2)))
 
-        Helper.append_to_csv(reports, self.target_folder.format("report.txt"))
-        Helper.append_to_csv(reports_rel, self.target_folder.format("reports.txt"))
-        # Helper.append_to_csv_special(specifies, self.target_folder.format("specifies.txt"))
-        self.db.querymany(self.s_query, specifies)
+        if self.first:
+            Helper.append_to_csv(reports, self.target_folder.format("report.txt"))
+            Helper.append_to_csv(reports_rel, self.target_folder.format("reports.txt"))
+            # Helper.append_to_csv_special(specifies, self.target_folder.format("specifies.txt"))
+        else:
+            self.db.querymany(self.s_query, specifies)
 
         logging.debug("Finished chunk {} in {}s".format(self.iterations, round(time.time() - self.time, 2)))
         self.time = time.time()
@@ -87,7 +92,7 @@ class ReportInserter:
             .replace(" ", "")
         a_id = self.addresses[a_key]
 
-        b_key = "".join(str(i) for i in [col['b_1'][2]]) \
+        b_key = "".join(str(i) for i in [col['b_1'][1]]) \
             .replace(" ", "")
         b_id = self.businesses[b_key]
